requirements.txt
asgiref==3.7.2
attrs==23.1.0
Django==4.2.7
django-admin-autocomplete-filter==0.7.1
django-admin-list-filter-dropdown==1.0.3
django-admin-rangefilter==0.11.2
django-debug-toolbar==4.2.0
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
drf-spectacular-sidecar==2023.10.1
factory-boy==3.3.0
Faker==20.0.3
inflection==0.5.1
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.12.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
uritemplate==4.1.1
